I did a LOT and now it's 7:20 pm...
Ok so I did the big plant and I am VERY happy about it! I also finished the curtain rod and the curtain. Got some help with the mirror leather thing too and finished that up and now working on the rug! Need to figure out how to add dirt to pots and add the actual mirror instead of the frame...